3.28.15 Comments, debugged CurrencyFragment DB integration, start billing integration

+++ app/src/main/java/com/helwigdev/a/dogecoinutilities/WalletSecondaryListener.java
@@ -5,5 +5,7 @@ package com.helwigdev.a.dogecoinutilities;
  * All code herein copyright Helwig Development 3/18/2015
  */
 public interface WalletSecondaryListener {
-	void onGetBFBalance(Float[] balances);
+	//define callback
+	//potentially could integrate this and WalletListener, but it would be a little messier than I'd prefer
+	void onGetBFBalance(Double[] balances);
 }
